Move genomics modules and split it into gene/variant_set_id
